Changed construtor serialbits type, was double should be long (ref qpid-861)
git-svn-id: https://svn.apache.org/repos/asf/incubator/qpid/trunk/qpid@641304 13f79535-47bb-0310-9956-ffa450edef68

+++ java/common/src/main/java/org/apache/qpid/util/Serial.java
@@ -13,7 +13,7 @@ public class Serial
     private long _max;
     private long _maxComparison;
 
-    public Serial(double serialbits)
+    public Serial(long serialbits)
     {
         if( serialbits < 2)
         {
